Separate tests into new test files
