+++ app/processing/parse-genesis-return-file.js
@@ -21,8 +21,10 @@ const parseGenesisReturnFile = async (csv, filename) => {
         // ledger: 'AP',
         filename
       }
+    } else {
+      return ''
     }
-  })
+  }).filter(x => x !== '')
 }
 
 module.exports = parseGenesisReturnFile
