+++ app/processing/parse-imps-return-file.js
@@ -24,8 +24,10 @@ const parseImpsReturnFile = async (csv, filename) => {
         // ledger: 'AP',
         filename
       }
+    } else {
+      return ''
     }
-  })
+  }).filter(x => x !== '')
 }
 
 module.exports = parseImpsReturnFile
